fix: updated ManipulationInput
Previously the input component could not be used with an object of a nullable type. Part of the problem was that the path property did not accept any property-chains. To solve this the given object-type (MainObj) is converted into a non-nullable type. This way all its properties may be accessed using the path property. To prevent null-pointer exceptions, when the object actually is null. The component returns null incase the object is null.

In doing so the `hidden` property was added to the ManipulationInput component so that fields that have no value can be set to hidden.