feat: created convertContactInformations module
This lib-module is used to create a customer-object based on the ContactInformations of a order (billing or shipping information).

While it converts the order-informations to a customer object it also parses the given address, by merging the two address fields of an order. It then converted the combined address into a tuple of street and houseNumber. It does so by looking at the address, when it includes a number, then one can assume that its a houseNumber, thus the address is then split into street and houseNumber.