feat: updated RegistrationProvider props
Added the following states to the RegistrationProvider:
- order
- baseInformation

The order-state keeps track of the current order that is being edited in the slide-over. The baseInformation-state keeps track of whose ContactInformation's were used for the default-values of the customer-object.

This way, the user can swap between the billing and shipping ContactInformations, when they different.