refactor: removed useEffects in RegistrationProvider
The idea behind the previous change was, that a user only sets the order-state. Then the customer-state get automatically updated based on the baseInformation-state. In other words, once the order-state is set the customer-object is set to either the billing or shipping Informations based on the baseInformation state.

However, doing it this, causes the RegistrationProvider to re-render. In doing so the customer-state is not updated correctly. e.g. billing -> shipping:

initial: customer-billing -> customer-shipping -> (because of re-render) customer-billing