feat: added Warning-Menu component
This component is rendered when the billing and shipping informations of an order are different. It then shows a warning and two buttons, that allow the user to select which one of the information are to be used.

To ensure that the CustomerProps component is re-rendered when the base-values are changed. For example from billing to shipping or the other way around. The key property of the CustomerProps component was set to be a string of order.id and baseInformation. This way the CustomerProps component gets updated when the baseInformations or oder.id is changed.