style: modified DisplayOrder visibility styles
Unified the min-width classes for both the header and value fields. Also set the min-width property for the date-fields and header.

Fixed a mistake in different paddings, where the headers had a horizontal padding of 4px and the property fields of 3px. Which resulted in the use of margins to re-align the headers.