feat: updated DisplayHeaders component
Added a className property so that the (now standalone) component can be adjusted from the outside.

Removed the margin as this component can now be used on its own. In other words it must not be used inside of the same (flex) container, to achieve the same size, as long as it has the same paddings, margins and so on.

The margin was previously used to account for the gap between the different order-elements and header-component.