feat: created FeedTitle component
This component renders the title and description of a feed. Using its properties one can provide an icon that is then rendered before the title. When the link property is provided, then the title and description is wrapped in a Link component that navigates to the given link, when clicked.