refactor: updated CustomerBuyHistory component
Instead of passing a customer object to the component, one now passes down the invoices that are to be displayed. The reason for this is, that this component is therefore more generic and can be used in different scenarios.

To use this component to display the customer history, one simply filters the invoices by their customer to match the customer in question.