feat: created ArticlesBuyHistoryPage
This page renders the history for a given set of articles grouped by the customer that bought the (some) of the given articles.

The requested ids are provided by the dynamic-route segment referred to as `ids`. When multiple articles are selected, then the ids are to be separated by a comma.