refactor: exported `dymanic = 'force-dynamic` in routes
Exported the const dynamic with the value 'force-dynamic'. By doing this the route it self does not cache any results. Which means that whenever the route is called the code within it actually executed. In other words each time the route is opened it will retrieve a set of articles, invoices, categories.

This can be done since each page that uses these endpoints cache the results-anyway.