feat: created ArticleTable component
This component renders a given array of articles in the form of a table. The table it self is responsive, as the visibility of its column depends on the container-size, thus the width of the table.

It features a simple action-button that appears when at least one article is selected. When clicked it shows the article-based-history for the selected articles.

The ArticleTable component also features a simple search-bar that filters the given articles.