refactor: modified TableProps interface
Removed the partial from the labels property. Instead not each key must be provided, thus can be either of type string or undefined.

At the same time the labels argument is no longer optional for the TableContext. The reason for this is, that this argument will receive default values, which makes them no longer optional properties.