feat: externalized the rendering of table-column-labels
The new `TableColumnLabels` function now renders the labels of the generic table in the table-header. It does this by iterating through the keys of the labels object and sets the table-header to the objects value of each key. Note, that the label object specifies a label for each key, given by the defaultLabels. That means, that when no labels are given to the table-component then, then the value of each property is the key it self.