feat: modified parent classes for children - root layout
Changed the structure for rendering the children elements in the root-layout.

The reason for this change is that, the container that wrapped the children did not "stretch" / cover the remaining space, at least its height didn't stretch.

Therefore the parent-div classes were changed to be fixed and the top, left properties were set to fill up the remaining space.

In order to provide a relative element for all the pages that are displayed a div element was placed inside of the fixed-container that wraps the children elements. Again the reason for this nested relative element is that elements inside of a page can placed with position absolute within the pages dimensions. In order to not loose the "stretching" behavior the relative-container has the w-full and h-full classes.

This way, the contents of the pages could take up the entire page-space using h-full.