fix: fixed missing scrollbars on overflow- root layout
Due to the recent change of wrapping the children elements of the root-layout inside a fixed container led to the fact that no scrollbars appeared when the content exceed the viewport-height. Therefore, the classes of the inner relative-container were updated to display vertical scrollbars automatically using `overflow-y-auto`.

Additionally, removed redundant classes: 'flex flex-col'
and moved padding to the inner-container.