feat: created `InvoiceHistoryFeed` component
Created a more generalized invoice-feed component. It generates a feed using the existing `Feed` components for the given invoices array.

By default this component will render a separate feed below an invoice showing its items.

This component also allows the customization of the description of the main feed for each invoice, as well as the icon shown in the main feed.