refactor: updated `/article-buy-history` page
Improved the code reliability by making sure to handle the case where an invoice does not have a customer object. In other words, the groupedCustomers function that maps the invoices of a customer to the tuple composed of his first and lastname or "Unknown Customer" in case the invoice has no customer.

The same thing has been done for the internal CustomerHistory component that shows the a heading about he invoice-history-feed with the customers-name. In case the reference invoice that was mapped to a customers-name does not exist, that means the customer-name would be "Unknown Customer", the heading displays "Unknown Customer".

Overall, these changes increase the robustness of this component, incase the provided invoices are not filtered as expected.