feat: created levelOnlyFormat helper module
This function can be used as a Format element in the configuration of winston. In other words, it can be used as any regular format element of winston inside the `combine` function.

This format will only "allow" / pass through a given level. That means, that the messages / logs of other levels are not going to be logged, when this format is applied. This can be useful, when the logs are saved in an file and a specific level that is somewhere in middle shall be saved in a file without any other levels.