feat: created createFileTransport helper module
This module exports a single function ´createFileTransport` which creates a winston-file-transport.

With the help of this function it is simple to create a new transport, that saves the logs in a given directory for a given level, where the filename is level.log. It also allows for customization, like whether or not the file is stored in a directory, whether it saves only one specific level, whether the file-transport is disabled and finally whether the file-name that is set by the function (level.log), shall be overridden with a custom filename.