feat: created `Logger` module
This module creates a new winston-logger and sets up the file-transports and console-transport. It then also export a few logging functions for each log-level that can be used both on the server and client-side.
However, these log-functions cannot be called on the initial render of a client-side component (throws an error).

The following logging-functions are exported:
- logError
- logWarning
- logInfo
- logDebug
- logVerbose