feat: drafted /invoices page
The newly created page shows a list of invoices using the generic Table component. Note, that the array of invoices has been modified, so that each element includes further properties that are used by the Table. Each table-row has a ItemButton that was previously inserted and redirects the user to the pdf-viewer page for that given invoice.