fix: drafted fix by reducing max staggering-delay
When a table displays a lot of items, thus shows a lot of rows, then the total amount of animation also increases. In that case the performance is drastically impaired until the animation has finished. However, even then the performance is still slightly impaired due to the huge amount of rows.