feat: create multi-staged Dockerfile
Using this Dockerfile the container will first install the necessary packages and then builds the nextjs-application. Note that the first step  also copies the .env file into the container to ensure a successful build. The final stage then copies the build-files, packages and other necessary files from the first stage. Its important to note that the final step does not copy the .env file and thus does not expose it.