feat: created Environment module
This module uses zod to verify that all the requested environment variables are present. It does this by building the zod-schema dynamically based on an array of properties. Then for each property it defines a custom error-message in case it is missing. Finally, it validates the enviroment variables and exports the verified env-variables.