feat: added Environment function call to layout
Now calling an empty function that is being exported by the Environment helper module, such that it is imported and therefore ensuring its top-level statements / validations are executed. This way we can ensure that the environment variables are validated on the first visit, and not just when it is used somewhere down the line.